The stripe API has been moved to features/store and the API route name has been changed.
