+++ src/features/webhook/stripe/type.ts
@@ -1,0 +1,3 @@
+export interface StripeWebhookResponse {
+	received: true;
+}
